+++ app/src/main/kotlin/io/sweers/catchup/ui/controllers/PagerController.kt
@@ -22,6 +22,7 @@ import android.animation.ArgbEvaluator
 import android.animation.ValueAnimator
 import android.content.Context
 import android.content.Intent
+import android.content.SharedPreferences
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.os.Parcelable
@@ -49,8 +50,10 @@ import com.bluelinelabs.conductor.RouterTransaction
 import com.bluelinelabs.conductor.support.RouterPagerAdapter
 import com.jakewharton.rxbinding2.support.design.widget.RxAppBarLayout
 import com.uber.autodispose.kotlin.autoDisposable
+import dagger.Provides
 import dagger.Subcomponent
 import dagger.android.AndroidInjector
+import io.sweers.catchup.P
 import io.sweers.catchup.R
 import io.sweers.catchup.changes.ChangelogHelper
 import io.sweers.catchup.injection.ConductorInjection
@@ -66,18 +69,16 @@ import io.sweers.catchup.util.isInNightMode
 import io.sweers.catchup.util.resolveAttributeColor
 import io.sweers.catchup.util.setLightStatusBar
 import io.sweers.catchup.util.updateNavBarColor
-import java.util.Arrays
 import javax.inject.Inject
 
-// TODO Temporary till static services are removed
-fun ServiceMeta.toStaticService() = StaticService(
+fun ServiceMeta.toServiceHandler() = ServiceHandler(
     name,
     icon,
     themeColor,
     { ServiceController.newInstance(id) }
 )
 
-data class StaticService(@StringRes val name: Int,
+data class ServiceHandler(@StringRes val name: Int,
     @DrawableRes val icon: Int,
     @ColorRes val accent: Int,
     val instantiator: () -> Controller)
@@ -88,54 +89,58 @@ class PagerController : ButterKnifeController {
 
     private const val SETTINGS_ACTIVITY_REQUEST = 100
     private const val PAGE_TAG = "PagerController.pageTag"
-    private lateinit var services: Array<StaticService>
   }
 
-  private lateinit var resolvedColorCache: IntArray
-  private val argbEvaluator = ArgbEvaluator()
+  @Inject
+  lateinit var resolvedColorCache: IntArray
+  @Inject
+  lateinit var serviceHandlers: Array<ServiceHandler>
+  @Inject
+  lateinit var changelogHelper: ChangelogHelper
 
-  @Inject lateinit var serviceMetas: Map<String, @JvmSuppressWildcards ServiceMeta>
-  @Inject lateinit var changelogHelper: ChangelogHelper
-  @BindView(R.id.pager_controller_root) lateinit var rootLayout: CoordinatorLayout
-  @BindView(R.id.tab_layout) lateinit var tabLayout: TabLayout
-  @BindView(R.id.view_pager) lateinit var viewPager: ViewPager
-  @BindView(R.id.toolbar) lateinit var toolbar: Toolbar
-  @BindView(R.id.appbarlayout) lateinit var appBarLayout: AppBarLayout
+  @BindView(R.id.pager_controller_root)
+  lateinit var rootLayout: CoordinatorLayout
+  @BindView(R.id.tab_layout)
+  lateinit var tabLayout: TabLayout
+  @BindView(R.id.view_pager)
+  lateinit var viewPager: ViewPager
+  @BindView(R.id.toolbar)
+  lateinit var toolbar: Toolbar
+  @BindView(R.id.appbarlayout)
+  lateinit var appBarLayout: AppBarLayout
+
+  // Not injectable because I don't know how to get a @BindsInstance into dagger's AndroidInjector
+  // and this needs a controller to be instantiated
+  private lateinit var pagerAdapter: RouterPagerAdapter
+
+  private val argbEvaluator = ArgbEvaluator()
   private var statusBarColorAnimator: ValueAnimator? = null
   private var tabLayoutColorAnimator: Animator? = null
   private var tabLayoutIsPinned = false
   private var canAnimateColor = true
   private var lastPosition = 0
-  private var pagerAdapter: RouterPagerAdapter
 
   constructor() : super()
 
   @Suppress("unused")
   constructor(args: Bundle) : super(args)
 
-  init {
+  override fun onContextAvailable(context: Context) {
+    ConductorInjection.inject(this)
+    super.onContextAvailable(context)
+
     pagerAdapter = object : RouterPagerAdapter(this) {
       override fun configureRouter(router: Router, position: Int) {
         if (!router.hasRootController()) {
-          router.setRoot(RouterTransaction.with(services[position].instantiator())
-              .tag(PAGE_TAG))
+          router.setRoot(RouterTransaction.with(serviceHandlers[position].instantiator())
+              .tag("$PAGE_TAG.${serviceHandlers[position].name}"))
         }
       }
 
-      override fun getCount() = services.size
+      override fun getCount() = serviceHandlers.size
 
       override fun getPageTitle(position: Int) = ""
     }
-  }
-
-  override fun onContextAvailable(context: Context) {
-    ConductorInjection.inject(this)
-    services = (serviceMetas.values.map { it.toStaticService() }).toTypedArray()
-
-    resolvedColorCache = IntArray(services.size)
-    // Invalidate the color cache up front
-    Arrays.fill(resolvedColorCache, R.color.no_color)
-    super.onContextAvailable(context)
   }
 
   override fun onSaveViewState(view: View, outState: Bundle) {
@@ -245,7 +250,7 @@ class PagerController : ButterKnifeController {
     }
 
     // Initial title
-    toolbar.title = resources!!.getString(services[0].name)
+    toolbar.title = resources!!.getString(serviceHandlers[0].name)
 
     // Set the initial color
     @ColorInt val initialColor = getAndSaveColor(0)
@@ -257,8 +262,8 @@ class PagerController : ButterKnifeController {
     }
 
     // Set icons
-    for (i in services.indices) {
-      val service = services[i]
+    for (i in serviceHandlers.indices) {
+      val service = serviceHandlers[i]
       val d = VectorDrawableCompat.create(resources!!, service.icon, null)
       tabLayout.getTabAt(i)!!.icon = d
     }
@@ -268,12 +273,12 @@ class PagerController : ButterKnifeController {
     viewPager.addOnPageChangeListener(object : ViewPager.OnPageChangeListener {
       override fun onPageScrolled(position: Int, positionOffset: Float, positionOffsetPixels: Int) {
         if (canAnimateColor) {
-          val color: Int = if (position < pagerAdapter.count - 1 && position < services.size - 1) {
+          val color: Int = if (position < pagerAdapter.count - 1 && position < serviceHandlers.size - 1) {
             argbEvaluator.evaluate(positionOffset,
                 getAndSaveColor(position),
                 getAndSaveColor(position + 1)) as Int
           } else {
-            getAndSaveColor(services.size - 1)
+            getAndSaveColor(serviceHandlers.size - 1)
           }
           tabLayout.setBackgroundColor(color)
           if (tabLayoutIsPinned) {
@@ -296,7 +301,7 @@ class PagerController : ButterKnifeController {
     tabLayout.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
       override fun onTabSelected(tab: TabLayout.Tab) {
         val position = tab.position
-        toolbar.setTitle(services[position].name)
+        toolbar.setTitle(serviceHandlers[position].name)
 
         // If we're switching between more than one page, we just want to manually set the color
         // once rather than let the usual page scroll logic cycle through all the colors in a weird
@@ -343,19 +348,20 @@ class PagerController : ButterKnifeController {
 
       override fun onTabReselected(tab: TabLayout.Tab) {
         pagerAdapter.getRouter(tab.position)?.getControllerWithTag(PAGE_TAG)?.let {
-          if (it is Scrollable) {
-            it.onRequestScrollToTop()
-            appBarLayout.setExpanded(true, true)
-          }
-        }
+              if (it is Scrollable) {
+                it.onRequestScrollToTop()
+                appBarLayout.setExpanded(true, true)
+              }
+            }
       }
     })
   }
 
-  @ColorInt private fun getAndSaveColor(position: Int): Int {
+  @ColorInt
+  private fun getAndSaveColor(position: Int): Int {
     if (resolvedColorCache[position] == R.color.no_color) {
       resolvedColorCache[position] = ContextCompat.getColor(dayOnlyContext!!,
-          services[position].accent)
+          serviceHandlers[position].accent)
     }
     return resolvedColorCache[position]
   }
@@ -365,7 +371,11 @@ class PagerController : ButterKnifeController {
     if (requestCode == SETTINGS_ACTIVITY_REQUEST) {
       if (resultCode == SettingsActivity.SETTINGS_RESULT_DATA && data != null) {
         val extras = data.extras
-        if (extras.getBoolean(SettingsActivity.NIGHT_MODE_UPDATED, false)) {
+        if (extras.getBoolean(SettingsActivity.NIGHT_MODE_UPDATED, false)
+            || extras.getBoolean(SettingsActivity.SERVICE_ORDER_UPDATED, false)) {
+          for (i in 0 until tabLayout.tabCount) {
+            pagerAdapter.getRouter(i)?.setBackstack(emptyList(), null)
+          }
           activity?.recreate()
         }
         if (extras.getBoolean(SettingsActivity.NAV_COLOR_UPDATED, false)) {
@@ -378,10 +388,30 @@ class PagerController : ButterKnifeController {
   }
 
   @PerController
-  @Subcomponent
+  @Subcomponent(modules = [Module::class])
   interface Component : AndroidInjector<PagerController> {
 
     @Subcomponent.Builder
     abstract class Builder : AndroidInjector.Builder<PagerController>()
+  }
+
+  @dagger.Module
+  object Module {
+
+    @JvmStatic
+    @Provides
+    fun provideServiceHandlers(sharedPrefs: SharedPreferences,
+        serviceMetas: Map<String, @JvmSuppressWildcards ServiceMeta>): Array<ServiceHandler> {
+      val currentOrder = sharedPrefs.getString(P.ServicesOrder.KEY, null)?.split(",") ?: emptyList()
+      return (serviceMetas.values
+          .sortedBy { currentOrder.indexOf(it.id) }
+          .map { it.toServiceHandler() })
+          .toTypedArray()
+    }
+
+    @JvmStatic
+    @Provides
+    fun provideColorCache(serviceHandlers: Array<ServiceHandler>) = IntArray(
+        serviceHandlers.size).apply { fill(R.color.no_color) }
   }
 }