+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* **New:** You can now reorder services in settings
 * **Enhancement:** GitHub is now paginated when there are more results (thanks to [@charlesdurham](https://github.com/charlesdurham)!)
 
 ## 0.2.2 (2017-12-17)